requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
PyMuPDF==1.23.7
pdfplumber==0.10.3
python-docx==1.1.0
pandas==2.1.4
openpyxl==3.1.2
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
numpy==1.26.2
scikit-learn==1.3.2
reportlab==4.0.8
Pillow==10.1.0
aiofiles==23.2.1
